refactor: 파일 구조 및 변수명 변경
