--- backend/src/main/java/food/backend/oauth/LoginParams.java
@@ -1,6 +1,0 @@
-package food.backend.oauth;
-
-public interface LoginParams {
-
-    public OAuthType getOAuthType();
-}
