+++ backend/src/main/java/food/backend/oauth/common/LoginParams.java
@@ -1,0 +1,6 @@
+package food.backend.oauth.common;
+
+public interface LoginParams {
+
+    OAuthType getOAuthType();
+}
