+++ backend/src/main/java/food/backend/oauth/common/OAuthType.java
@@ -1,4 +1,4 @@
-package food.backend.oauth;
+package food.backend.oauth.common;
 
 public enum OAuthType {
     KAKAO, NAVER, GOOGLE
