+++ backend/src/main/java/food/backend/oauth/common/TokenInfo.java
@@ -1,10 +1,12 @@
-package food.backend.oauth;
+package food.backend.oauth.common;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
+@Builder
 public class TokenInfo {
     private String accessToken;
     private String refreshToken;
