+++ backend/src/main/java/food/backend/oauth/kakao/entity/KakaoToken.java
@@ -1,7 +1,7 @@
-package food.backend.oauth.entity;
+package food.backend.oauth.kakao.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import food.backend.oauth.TokenInfo;
+import food.backend.oauth.common.TokenInfo;
 import lombok.Getter;
 
 @Getter
@@ -24,7 +24,4 @@ public class KakaoToken {
     @JsonProperty("scope")
     private String scope;
 
-    public TokenInfo of() {
-        return new TokenInfo(accessToken, refreshToken);
-    }
 }
