+++ backend/src/main/java/food/backend/oauth/kakao/entity/KakaoUserInfo.java
@@ -1,0 +1,13 @@
+package food.backend.oauth.kakao.entity;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+
+@Getter
+public class KakaoUserInfo {
+
+    @JsonProperty("id")
+    private Long memberId;
+
+
+}
